+++ Basil/include/Core/Log.h
@@ -5,6 +5,7 @@
 
 #include "Core.h"
 
+#include <glm/gtx/string_cast.hpp>
 #include <spdlog/spdlog.h>
 #include <spdlog/fmt/ostr.h>
 #include <spdlog/sinks/basic_file_sink.h>
@@ -23,6 +24,26 @@ namespace Basil
 			static Shared<spdlog::logger> engineLogger;
 	};
 }
+
+// Allow logging of GLM vectors, matrices, and quaternions
+template<typename OStream, glm::length_t L, typename T, glm::qualifier Q>
+inline OStream& operator<<(OStream& os, const glm::vec<L, T, Q>& vector)
+{
+	return os << glm::to_string(vector);
+}
+
+template<typename OStream, glm::length_t C, glm::length_t R, typename T, glm::qualifier Q>
+inline OStream& operator<<(OStream& os, const glm::mat<C, R, T, Q>& matrix)
+{
+	return os << glm::to_string(matrix);
+}
+
+template<typename OStream, typename T, glm::qualifier Q>
+inline OStream& operator<<(OStream& os, const glm::qua<T, Q>& quaternion)
+{
+	return os << glm::to_string(quaternion);
+}
+
 
 // Define client logging macros
 #define LOG_CLIENT_TRACE(...)	Basil::Log::getClientLogger()->trace(__VA_ARGS__);